LayoutTests/scrollingcoordinator/mac/latching/zero-delta-began-should-not-latch-expected.txt

Tests a gesture with zero delta on the began event doesn't latch on the document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS overflow.scrollTop is 0
PASS window.pageYOffset is 0
After scroll
PASS overflow.scrollTop is 200
PASS window.pageYOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
